requirements.txt
docopt==0.6.2
et-xmlfile==1.1.0
numpy==1.23.4
openpyxl==3.0.10
pandas==1.5.1
python-dateutil==2.8.2
pytz==2022.6
six==1.16.0
webvtt-py==0.4.6
XlsxWriter==3.0.3
